PlayCookieStore (#166)
* Update pom.xml

* Create PlayCookieStore.java

Initial version

* Javadocs, clean constructor

updated for according to the discussion in pac4j/play-pac4j#166

It wasn't clear if I should remove the cache, so I haven't yet. Doing so upon request should be easy.

* Remove optional cache

Per the discussion in pac4j/playpac4j#166

* PlayCookieStore Instructions

* Remove unnecessary code

Also added a check to ensure that the encryptConfig is not null

* Comparison of PlayCookieStore with PlayCacheStore