Replace deprecated methods
Replaced the `getHeader` method with the recommended `header` method, which returns an optional (this is somewhat related to #239, since this deprecated method will be removed in the next Play version).
Replaced the `IOUtils.closeQuietly` method with a `try-with-resources` structure.
As an added bonus, the `compressBytes` method now also closes its `OutputStream` when an exception occurs.