Increase security in ShiroAesDataEncrypter (#273)
Deprecate the constructor which takes the AES key as a String.  The
mapping from String values into the AES key space was not surjective.

Generate more secure random keys in the default constructor: sample
from the entire key space, and use a SecureRandom to do it.

Add a constructor which takes the AES key as a byte array, enabling
the user to supply their own key.